+++ jdaviz/configs/cubeviz/plugins/tests/test_parsers.py
@@ -150,20 +150,25 @@ def test_spectrum3d_parse(image_cube_hdu_obj, cubeviz_helper):
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_fits_image_hdu_parse_with_inverse_var(image_cube_hdu_obj, deconfigged_helper):
+@pytest.mark.parametrize(
+    ('unc_extname', 'expected_stddev_value'),
+    [('IVAR', 0.5), ('VAR', 2.0)]
+)
+def test_fits_image_hdu_parse_with_inverse_var(image_cube_hdu_obj, deconfigged_helper,
+                                               unc_extname, expected_stddev_value):
     """
-    Test loading a cube that contains an IVAR (inverse variance) extension for
-    its uncertainty. the IVAR extension will be converted to a standard deviation
-    type internally for consistency.
+    Test loading cubes that contain IVAR (inverse variance) and VAR (variance)
+    uncertainty extensions. Both should be converted to standard deviation
+    internally for consistency.
     """
 
-    # use existing test fixture, but change ERR ext to IVAR
+    # use existing test fixture, but change ERR ext to requested uncertainty type
     hdul = image_cube_hdu_obj.copy()
-    hdul[2].name = 'IVAR'
-    hdul[2].header['EXTNAME'] = 'IVAR'
+    hdul[2].name = unc_extname
+    hdul[2].header['EXTNAME'] = unc_extname
 
-    # make the values 4.0 (which, when converted to StdDevUncertainty, should be
-    # 0.5 which we'll check later)
+    # with values initially set t\o 4.0, when converted from IVAR to stddev,
+    # we should get 0.5 and when converted from VAR to stddev, we should get 2.0
     hdul[2].data = np.full_like(hdul[2].data, 4.0)
 
     deconfigged_helper.load(hdul, format='3D Spectrum')
@@ -171,14 +176,13 @@ def test_fits_image_hdu_parse_with_inverse_var(image_cube_hdu_obj, deconfigged_h
     # data collection should contain flux, mask, uncert, and extracted spectrum
     assert len(deconfigged_helper.app.data_collection) == 4
 
-    # now check the actual values of the uncertainty, since they were 4.0 in inverse
-    # variance they should be 0.5 in standard deviation
+    # now check the actual values of the uncertainty after conversion to stddev
     unc_data = deconfigged_helper.datasets['3D Spectrum [UNC]'].get_data()
 
     # the BUNIT keyword in the header is applying a factor of
     # '1E-17 erg*s^-1*cm^-2*Angstrom^-1' so account for that in the expected value here
     bunit = 1E-17
-    assert_allclose(unc_data.flux, 0.5 * bunit * unc_data.unit)
+    assert_allclose(unc_data.flux, expected_stddev_value * bunit * unc_data.unit)
 
 
 @pytest.mark.parametrize("flux_unit", [u.nJy, u.DN, u.DN / u.s])